+++ pom.xml
@@ -140,7 +140,6 @@
         <module>spark-java</module>
       <!--  <module>spring-5</module>-->
         <module>spring-5-mvc</module>
-        <module>spring-acl</module>
         <module>spring-activiti</module>
         <module>spring-akka</module>
         <module>spring-amqp</module>
@@ -199,6 +198,7 @@
         <module>spring-rest-query-language</module>
         <module>spring-rest</module>
         <module>spring-rest-simple</module>
+        <module>spring-security-acl</module>
         <module>spring-security-cache-control</module>
         <module>spring-security-client/spring-security-jsp-authentication</module>
         <module>spring-security-client/spring-security-jsp-authorize</module>
