+++ spring-security-acl/pom.xml
@@ -4,7 +4,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>com.baeldung</groupId>
-    <artifactId>spring-acl</artifactId>
+    <artifactId>spring-security-acl</artifactId>
     <version>0.0.1-SNAPSHOT</version>
     <packaging>war</packaging>
